+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/AutoAdaptableKubernetesClient.java
@@ -64,6 +64,7 @@ import io.fabric8.kubernetes.api.model.Service;
 import io.fabric8.kubernetes.api.model.ServiceAccount;
 import io.fabric8.kubernetes.api.model.ServiceAccountList;
 import io.fabric8.kubernetes.api.model.ServiceList;
+import io.fabric8.kubernetes.client.utils.PodGeneratorImpl;
 import okhttp3.OkHttpClient;
 
 import java.io.InputStream;
@@ -259,6 +260,11 @@ public class AutoAdaptableKubernetesClient extends DefaultKubernetesClient {
   @Override
   public MixedOperation<LimitRange, LimitRangeList, DoneableLimitRange, Resource<LimitRange, DoneableLimitRange>> limitRanges() {
     return delegate.limitRanges();
+  }
+
+  @Override
+  public PodGeneratorImpl run() {
+    return delegate.run();
   }
 
   @Override